helpers/widget: A simple helper to easily add widgets to a page (template)
This helpers makes it really easy to add a new widget to a page.

Example:
$widget = widget::add('netw_health', array('index'));
The first argument is the name of the widget we want to include and the
second argument is an array of options passed to the widget controller.
The first parameter in the second argument ('index' above) is the method
that will be called in the widget controller.

Signed-off-by: Per Åsberg <per.asberg@op5.se>