Fire_Profiler: add FirePHP options to Fire_profiler
Somehow the options for FirePHP are lost when creating a new instance
with init() from Fire_Profiler. It's no big issue and we don't depend
on it anyway but using it might be convenient sometimes when we want
profiling without clogging up the page output.
Copying the options array into Fire_Profiler::__construct() makes it
work again.

Signed-off-by: Per Åsberg <per.asberg@op5.se>