models/current_status: Introducing a PHP version of TAC
This is the first stab at creating a PHP version of tac.c
The code is basically ported from C to PHP with the use of database
storage for status data instead of file storage.
It is currently totally untested due to the lack of status data in
the database. There is however get_network_health() which uses
get_nagios_status() to parse status.log to find data that the netw_health
widget needs to work.

Signed-off-by: Per Åsberg <per.asberg@op5.se>