controllers: Let base class handle view paths
Since we need to be able to point our views to different directories
depending on current theme and configuration, we now let the base (ninja)
class handle the paths for us. Instead of using new View(), which is the
kohana default, we now use $this->add_view(). this way we will never
have to worry about where the current theme is located.

Signed-off-by: Per Åsberg <per.asberg@op5.se>