Use encode/decode for host/service link parameters
Sending a service name parameter such as 'Disk usage /' will break
Kohanas parameter handling as it seems to assume that it is a parameter
delimiter. Urlencoding or even rawurlencoding won't help here.
The solution this far is to create a helper class (link) to handle it for us.
Calling link::encode() with a string will base64 encode and rawurlencode
it for us to make it safe to pass to a controller/method.

The drawback on this is that it will be impossible to create a url to
a controller without using the link helper class but I honestly can't
come up with anything else that seems to work properly.

Signed-off-by: Per Åsberg <per.asberg@op5.se>