controller/status/convert_status_value: handle "All Problems" and types
The convert_status_value() method is used to convert state values from
status_totals (which uses the same state values as Nagios) to something
that we can use for querying the database. This patch makes it special
case values that are used for filtering out 'All Problems' or 'All Types'.

Signed-off-by: Per Åsberg <per.asberg@op5.se>