controllers/status/service: Added controller method and view for service
New controller method and view created for services.
It takes care of the status service listing. This could have been created
in the host method as well but keeping them apart will make the logic
simpler and we avoid a lot of the if/elseif/else statements that we
otherwise would been doing. The drawback is probably that we will have
to maintain two different views that are pretty similar but this will
have to do for now.

Signed-off-by: Per Åsberg <per.asberg@op5.se>