widgets/status_totals: Added new widget - status_totals
This widget is a re-write of the Nagios filter table located top right
on status.c
It is supposed to behave like the original but it is now also
updated in background by using ajax calls. The refresh interval is
configurable but just now you have to edit the javascript file to
change the default (60s). The plan is to make this configurable when we
have enabled the edit-widget functionality.

Signed-off-by: Per Åsberg <per.asberg@op5.se>