Added new controller and view for extinfo
This is the first stab at the new extinfo controller to substitute
the old extinfo.cgi. It is currently capable of showing details for
host and services but the presentation is pretty raw and needs to be
worked on quite a bit. No css or anything included so even the old
extinfo.cgi seems pretty when compared :)

Signed-off-by: Per Åsberg <per.asberg@op5.se>