controllers/extinfo: Add default controller method
In Nagios the extinfo CGI is called with type=0 when displaying process
info. Instead of directly calling extinfo::show_process_info() we may
just call extinfo/ which redirects to show_process_info(). Not a very big
issue but it seems to be in the line of how Nagios works so it can't be
that bad. The future will tell.

Signed-off-by: Per Åsberg <per.asberg@op5.se>