widgets/status_totals: Pass widget arguments in ajax call
Since the status_totals widget should display data for a wide range
of circumstances, we need to figure out what we are currently showing
to be able fetch the correct data when doing the ajax update.
This is done by parsing the links in the widget and passing the correct
values back in the ajax call for the receiving PHP method to use.

Signed-off-by: Per Åsberg <per.asberg@op5.se>