views/status/service.php: Re-add modified view for status::services()
We removed part of this patch before since the translation into images
didn't seem like the way to as we believed it would require translators
to add new images. As it turns out, it doesn't show any translations but
rather uses a badly named method (fixed in previous commit) to decide
what image to show.

Signed-off-by: Per Åsberg <per.asberg@op5.se>