helpers/nagstat.php: Alter host and service state constants
In core nagios, service-states go as follows:
  SERVICE_OK = 0
  SERVICE_WARNING = 1
  SERVICE_CRITICAL = 2
  SERVICE_UNKNOWN = 3

There is no SERVICE_PENDING. But in the CGI's the ordering
of the service-states go like this:
  SERVICE_PENDING = 1
  SERVICE_OK = 2
  SERVICE_WARNING = 4
  SERVICE_UNKNOWN = 8
  SERVICE_CRITICAL = 16

In other words, UNKNOWN and CRITICAL are flipped in order,
and PENDING is inserted at the top of the list. This change
adds PENDING as a magic number 64, and re-orders the rest of
the macros so that they follow the ordering as it's given
in-core in nagios.

Signed-off-by: Andreas Ericsson <ae@op5.se>