Add httpd configuration for ninja
We don't want to use symlinks from the webroot if we can avoid
it, since that doesn't always play well with other systems'
directory hierarchy. Instead, we use an apache alias that will
cause the right directory to be loaded when the user points
his or her browser to the example.com/ninja url.

Signed-off-by: Andreas Ericsson <ae@op5.se>