widgets/status_totals.js: Fetch correct pathname
Since the status_totals widget is used on a lot of pages, yhe first
thing the ajax callback method needs to do is to figure out what
parameters to send back to the widget.
It does this by looking at the links on the page but this wasn't
very intelligent in previous versions since it would break as soon
as the page layout would change.
This patch uses JQuery to find the first link in the widget-host_totals
div. It shouldn't break unless we change this div name.

Signed-off-by: Per Åsberg <per.asberg@op5.se>