controllers/status: Refactor to enable support for hostgroups
By refactoring a bit, we get support for hostgroups as well as
servicegroups. servicgroup() was renamed to group(), with first argument
being the type of group, and added 2 wrappers for host- and servicegroups
that calls group with correct values.
show_servicegroup() was renamed the same way to _show_group() with
grouptype as first argument.

Signed-off-by: Per Åsberg <per.asberg@op5.se>