Implement and use a Ninja authentication driver
Since we need to support passwords hashed with a lot of different
algorithms, we need to use our own authentication driver for this.

Also, since Kohana's Auth module is utterly retarded and refuses
to pass the plaintext password to the driver, we hack our way
around that, calling the driver implicitly if we're using the
Ninja authentication driver (which is default as of now).

With this patch, we have support for:
 * sha1    : 40 char SHA-1 hash
 * b64_sha1: Base64 encoded binary SHA-1 of password
 * crypt   : Plain old "crypt(3)" encrypted password hash
 * plain   : Plaintext password (only for testing purposes)
 * apr_md5 : The apr library's convoluted implementation of
             the MD5 hash algorithm.

All we need now is a small script to import currently existing
users into the database with their existing (htpasswd) encrypted
passwords and everyone who previously had access to the user
interface will continue to have such access.

Signed-off-by: Andreas Ericsson <ae@op5.se>