widgets/status_totals: Return if no data
To prevent widgets from breaking the parent page just because of no
available data, we simply return. This should probably be handled
in a more sophisticated way by showing some information to the user
but one could argue that if the widget can't find any data, the parent
controller won't either and should therefore be responsible for
handling the case.

Signed-off-by: Per Åsberg <per.asberg@op5.se>