theme/host: Don't operate on possibly unset arrays (3rd time...)
To avoid unnecessary indentation, we create the array as an empty
one in case PHP considers it empty(). Best case scenario, the
code never triggers and we waste one cycle. Worst case scenario,
we enter the loop with an empty array, and exit it again immediately.

Without this patch, the worst case scenario involves Kohana bombing
out.

The code introduced in this patch was accidentally removed in commit
2910e86da28272f95ecf36926c98e71dbd0845a8, was cherry-picked back
later and then accidentally removed again in commit
dd9fd9a88044896502e8850be50672868ce171fd. This commit re-introduces
it for (hopefully) the last time, and now with a big fat comment on
top of the if() clause. If this does not help, I shall be forced to
bring out the LART...

Signed-off-by: Andreas Ericsson <ae@op5.se>