theme/host: Don't operate on possibly unset arrays (again)
To avoid unnecessary indentation, we create the array as an empty
one in case PHP considers it empty(). Best case scenario, the
code never triggers and we waste one cycle. Worst case scenario,
we enter the loop with an empty array, and exit it again immediately.

Without this patch, the worst case scenario involves Kohana bombing
out.

The code introduced in this patch was accidentally removed in commit
2910e86da28272f95ecf36926c98e71dbd0845a8, but it needs to be there
so we cherry-pick it back.

Signed-off-by: Andreas Ericsson <ae@op5.se>