theme/host: Don't operate on possibly unset arrays
To avoid unnecessary indentation, we create the array as an empty
one in case PHP considers it empty(). Best case scenario, the
code never triggers and we waste one cycle. Worst case scenario,
we enter the loop with an empty array, and exit it again immediately.

Without this patch, the worst case scenario involves Kohana bombing
out.

Signed-off-by: Andreas Ericsson <ae@op5.se>