host template: Check for empty()ness rather than array()ness
Apparently the $result variable is *not* an array but can still
be looped over by using foreach(). We check to see if it's
empty instead of checking if it's an array. This accomplishes
the same thing (it seems), but without fscking up the host
status page when there are actually valid results to show.

Signed-off-by: Andreas Ericsson <ae@op5.se>