models/current_status: Make status_text() handle groups
The method status_text() only handled host or serviceobjects and since
the status controller uses the return value from service() when displaying
host- and servicegroup details and the template uses router::$method as
argument to status_text(), we must handle this differently.
The easiest way is therefore to make status_text() handle servicegroups as
service and hostgroups as host.

Signed-off-by: Per Åsberg <per.asberg@op5.se>