models/service: Only return services on hosts in the selected hostgroup
A missing WHERE clause in the query for service-status regarding
services in a particular hostgroup previously caused us to return
a result-set containing num_hostgroup_members * num_services rows,
which naturally has the potential to quickly become unwieldy.

For the particular test-configuration used now, we previously
returned 635508 rows. With the fix, we get 254 rows instead,
which clearly is far more manageable.

Signed-off-by: Andreas Ericsson <ae@op5.se>