controllers/extinfo: Add pagination to comments listing
This is the first stab at trying to get a proper pagination in the
comments listing. It still has some quirks that needs to be worked out.
For one, it will in it's current state try to step in the pagination
for both host- and servicecomments if either one of them is clicked.
This will have to do for now.

Signed-off-by: Per Åsberg <per.asberg@op5.se>