made a temporary solution for sorting the service table
to be able to use the tablesorter script i decided to display the
host rows at all time. i've started on a script for using server-
side sorting, but that has a whole other type of sorting and i
either need some help with database information etc. or jquery
help to hide duplicate host-rows. at least the tables are sortable
now.

Signed-off-by: My Appelgren <my@op5.com>