models/host: Don't use IN() when user is authenticated for all services
If we always build the queries using IN() we will get huge queries which
will significantly slow down the application. This is the first patch to
try to amend this. It will only make a difference when user is authenticated
for all services and the next part should be to integrate the queries from
nagios_auth_class into all the queries without using IN().

Signed-off-by: Per Åsberg <per.asberg@op5.se>