controllers/search + template/result: Add global search
When a user enters something in the "Show host" field (which by the way
isn't a good label anymore), the query is sent to search::lookup() which
tries to find any object matching the query. It is also possible to narrow
down the search to a specific object type by entering it before the search
term. Ex: service:ping will search for all ping services available.

The result is displayed pretty much like status pages for host and services
and somewhat more rudimentary for groups.

Signed-off-by: Per Åsberg <per.asberg@op5.se>