controllers/status: Add num_per_page for paging
The nr of items displayed on a page where we use paging should be
configurable. So far there is no place in the GUI to do this but
by adding this GET variable we make the paging object honor it for both
host() and service() with a default of 10 items.

Signed-off-by: Per Åsberg <per.asberg@op5.se>