controllers/status: Enable paging for host() and service()
We use the built-in pagination class provided by Kohana to make
our listings more manageable. The down-side of this is that we always
have to make 2 queries for each listing - one to provide the total count
and one to fetch the actual content. This will however be a fair tradeoff
since the page loading gets much faster than showing everything when the
number of returned objects gets very big.

Signed-off-by: Per Åsberg <per.asberg@op5.se>