config/pagination: Added default pagination options
By editing this config file we save us some work when adding pagination
for pages. The options included here means that we use the 'digg' style,
pagination is not visible when the nr of returned items are less that the
num_per_page variable set and that we use the GET string 'page' to keep
track of what page to show.

Signed-off-by: Per Åsberg <per.asberg@op5.se>