helpers/json: Provide a JSON helper class for encode/decode
In our current PHP version we don't have built-in support for json_encode
or json_encode and have to instantiate the ZEND_Json class. Since the built-in
support will be there as soon as we reach the correct PHP version and as
we can't be sure what versions people will be running this on, we provide
a helper class that takes care of the checks of what json_{en, de}code
function to use.

Signed-off-by: Per Åsberg <per.asberg@op5.se>