models/nagios_auth: Modified authorized_service_query() method
It turned out that the queries that uses this method could be further
improved and it required some changes to the data returned.
The sprintf part was removed since we're better off joining the already
existing tables with the resulting query fields. To make this more
transparent to the calling method, we now also return the aliases of the
host and service tables.

Signed-off-by: Per Åsberg <per.asberg@op5.se>