controllers/authenticated + master template: Added public inline_js property
Sometimes we need to create dynamic javascript code that should be
inserted in the HTML head. This patch adds a public property for all
extending classes to use. It is intended to be of type string so
appending should work just fine.
The result is printed inside a JQuery $(document).ready part to be
executed once the page has finished loading.

Signed-off-by: Per Åsberg <per.asberg@op5.se>