tac_widgets.js: Reworked the javascript file to control the widgets for tac
Even though this script should be possible to use on almost any page that
hold widgets, it is currently only included in tac::index(). Since we
really don't have widgets anywhere else we can leave it like this for now.

All the cookie-handling-stuff has now been removed.

The widgets now support all the hide/close callbacks but we still have to
handle the edit callback and communicate this back to the actual widget.
Ah, well - this will do for now.

Signed-off-by: Per Åsberg <per.asberg@op5.se>