ninja_settings.sql: SQL statements needed for widgets <-> database
This script should be run against the merlin database during instalaltion
but until then we will have to run it manually.
It creates 2 new tables:
ninja_settings, which is intended to keep various page settings and
ninja_widgets, which holds the information about what widgets are available
globally or for a specific user for each page.

Signed-off-by: Per Åsberg <per.asberg@op5.se>