helpers/widget + ninja.sql: Assume widget method is index()
Since we currently are the only ones creating widgets we may change
the required parameters etc. This patch makes the assumption that all
widgets main class should be index(). This means that we don't have
to pass the method parameter for each widget and hence might skip all
default settings for the widgets saved in the database.

Signed-off-by: Per Åsberg <per.asberg@op5.se>