widgets/tac_problems: Add css to make slider work
It turns out that the JQuery UI slider needs some css to work properly.
This is just a minimal subset of the available styles and it could be that
this should be moved to a global stylesheet to save the widgets from having
to include it if a slider (or anything else from JQuery UI) is required.

Signed-off-by: Per Åsberg <per.asberg@op5.se>