widgets/tac_problems.js: Updated javascript to be able to save settings
Added some javascript logic to get the slider and to save settings
to the database. It uses SetInterval() + clearInterval() to control
the update interval and setTimeout() + clearTimeout()  to control
the settings update to database.
To prevent namespace collisons, all variables and methods are named
tac_problems_<something>. This is the desired behaviour for all widgets.

Signed-off-by: Per Åsberg <per.asberg@op5.se>