widget/tac_problems.js: Added comments and made update a generic function
It seems like a bad idea to have all the widgets re-inventing the wheel
every time they wants to refresh. This patch includes a generic way
of performing a refresh via AJAX call by using update_widget() with
the name of the widget and the ID of the widget view that should be
updated with returned content.

Signed-off-by: Per Åsberg <per.asberg@op5.se>