widget/tac_problems.js: Added in place editing of widget title
This is probably more of a proof of concept than something really
useful. It makes the widget title editable by double clicking and
saves the new string to the database for the current user.
There is, however, currently no way to restore the original name even
though this information should be easy to get from the database.

Signed-off-by: Per Åsberg <per.asberg@op5.se>