controllers/status: Replaced Kohana::config calls with config helper
Since we want to move some (but likely not all) of the config entries
from the config files to the database, we need a transparent way of
fetching these settings. By using the newly introduced config helper,
we can achieve this in a transparent and simple way.
The backside is that all Kohana::config() calls has to be replaced with
config::get(). This patch does just that for the status controller.

Signed-off-by: Per Åsberg <per.asberg@op5.se>