helpers/pnp: New helper class for PNP
This helper class takes care of things like fetching the PNP config
according the the pnp4nagios_config_path as setup in config/config.php

The config information is stored in the current session to prevent
unnecessary look-ups.
The has_graph() method should be called for each item where we want to
display a PNP link and it also replaces the current checks against
config.pnp4nagios_path.

Signed-off-by: Per Åsberg <per.asberg@op5.se>