widgets/tac_problems/view: Make widget header more generic
When enabling editing of the widget there are some IDs that should
(must) be present. Editing this info by hand is usually not a good
idea since a typo will break the functionality. By assigning the
widget_id to a variable in the controller, we can echo out this
to get consistency.

Signed-off-by: Per Åsberg <per.asberg@op5.se>