tac_widgets.js: Introducing our new widget javascript API
This patch contains a javascript class that makes it possible for all
the widgets to get a lot of functionality by just creating a new
widget instance. All AJAX update, slider for refresh editing and saving
of settings is performed from this class. It now only requires one line
of javascript in the widget js-file to get all this stuff (apart from
the HTML elements that needs to be available in the view).

Signed-off-by: Per Åsberg <per.asberg@op5.se>