removed float: left from form and moved it to the right of the pagination
i also noticed the custom pagination field doesn't work anymore, but it
seems it's caused by somthing else, and not my last changes. i'll commit
this as it's more or less the same as before yet a step ahead. perhaps
it's just a glitch and will work after i've synced with git..

Signed-off-by: My Appelgren <my@op5.com>