widgets: Create current_status instance for ajax calls
I'm not entirely sure that this is a good idea.
What the patch actually does is that it creates a new instance of the
current_status model and fetches all the data - again. For all the
widgets. This will probably increase the load on the server.
Consider this patch a temporary test that should be redesigned.

Signed-off-by: Per Åsberg <per.asberg@op5.se>