helpers/refresh: Added helper for page refresh
Since we want the refresh rate functions to be compatible with more than
the latest versions, we use an approach that overloads the same function
name. Since it's only the last version compatible with the current browser
that is used, we can get around the issue in a pretty nice way.
This script is printed directly to the browser and not included as the
other javascripts. This could probably be worked around but it will have
to suffice for now.

Signed-off-by: Per Åsberg <per.asberg@op5.se>