views/extinfo/index: Show info on contactgroups and contacts
This will print all info on all the contact groups that the current
object notifies to. It will also print info on what contacts that are
members of each contact group even though this info will be initially
hidden and the user will have to click on the name of the contact group
to reveal it.

Partial fix for issue #1757

Signed-off-by: Per Åsberg <per.asberg@op5.se>