helpers/refresh: Don't do refresh when user isn't logged in
It's rather silly to reload the login form and such and since the
refresh helper tries to find a page_refresh value from the database
for the current user it will fail instantly.
This patch makes the refresh helper to return if it detectst that
the user isn't logged in.

Signed-off-by: Per Åsberg <per.asberg@op5.se>