controllers/status: Avoid "Undefined index" runtime error
When we're finding service-problems it often makes sense to
only display the services not on problem-hosts. Previously
that particular bitmask selection lacked a translation,
causing Kohana to bomb out with an "Undefined index" warning.

This patch adds those translations both for services and hosts.
Although we do not use the one for services at the moment, it
we add it anyway for completeness.

Signed-off-by: Andreas Ericsson <ae@op5.se>