Set filter variables as class members
The filter variables hoststatustypes, servicestatustypes, hostprops
and serviceprops are passed around from calls to methods to helper
methods. This isn't a proper way of doing things when we have a class
instance that can carry the variables for us.
This patch adds these variables and assigns them in our models.

Signed-off-by: Per Åsberg <per.asberg@op5.se>