models/user: Added create_auth_table()
It might seem like an odd thing to do to create database tables from
within a script but it has the advantage that existing installations
can benefit from them directly. another advantage is that we can leave
it up to the application to decide what database driver to use which,
so far, is a rather awkward thing to do from install scripts.

This method won't ever be used on fresh
installations since the table will already be created by the install
script.

Signed-off-by: Per Åsberg <per.asberg@op5.se>