install_scripts/auth_import: Move functionality into Ninja
Since we already have a lot of code in Ninja to do what we want to
do during install, it seems like a bad idea to create it all over again.
By using the CLI possibilities introduced earlier, we can now call a method
in the cli controller to achieve what we want.

Signed-off-by: Per Åsberg <per.asberg@op5.se>