Add a utility to import passwords from htpasswd files
Since all our customers already have a lot of users configured
in their htpasswd files, we need a way to import those into the
ninja database. This script provides a means to do just that,
and is a must-have in order to deliver Ninja to op5 customers.

The authentication driver was written with htpasswd passwords
in mind, so this plays nicely with how Ninja works today.

Signed-off-by: Andreas Ericsson <ae@op5.se>