controllers/default: Added methods for CLI access
When we are accessing ninja from commandline we want to know what the
current cli_access settings are before we try to do anything. If it is set
to false, we can't do anything and if set to true we need to find any user
with a login role. This patch contains the methods needed for getting this
data. Since they are created in the default controller, they don't require
any authentication but since they are only accessible from command line
we can feel safe anyway.

Signed-off-by: Per Åsberg <per.asberg@op5.se>