controllers/command: Allow users to force scheduled checks
This adds basic exception handling to the command controller
that makes it possible to add and extract random variables
from the command input parameters. The first exception we add
is to create a "Force Check" checkbox button (which defaults
to "on") for all the various ways there are for scheduling
checks.

This fixes issue #2504.

Signed-off-by: Andreas Ericsson <ae@op5.se>