command templates: s/;host_name;service_description;/;service;/
Since the way we list services turn out to be identical when we
use it to replace the ';service;' string in the template anyway,
we might as well take this step to be 100% certain of when we're
looking for a host_name;service_description combination.

This also makes the $remove array quite useless in the command
model, since nothing uses it any longer.

Signed-off-by: Andreas Ericsson <ae@op5.se>