controllers/command: Support scheduling downtime for services of a host
When scheduling downtime for a host, it's normal (and usual) to
have that host go offline during the downtime period. The hosts
services will then go into a non-ok state, and will therefore
alert. This patch adds a small checkbox to the two commands that
schedule downtime for one or more hosts. If the checkbox is
checked (the default for it), downtime will be scheduled for all
services attached to the host, or to all hosts in the hostgroup.

This fixes issue #767.

Signed-off-by: Andreas Ericsson <ae@op5.se>