controllers/command: Refactor creating checkbox variables
Since we're primarily using checkbox variables and most of
the options we set for those variables are identical, we
can get rid of quite a lot of text by simply using a small
wrapper to create the checkbox variable arrays for us.

We also cut down the line length somewhat by pulling out the
$info['params'] array to a variable of its own and then
re-assigning it into $info['params'] when we're done.

Signed-off-by: Andreas Ericsson <ae@op5.se>