controllers/ajax: Add ajax detection variable
Since some browsers (IE7 for one) doesn't seem to supply us with the
HTTP_X_REQUESTED_WITH header information, we can't always rely on this.
This patch adds the is_ajax variable to the arguments array passed to
all widgets and makes it possible for them to unambiguously detect if
the call is coming from the ajax::widget controller.

Signed-off-by: Per Åsberg <per.asberg@op5.se>