fixed strange ie7 bug for status pages
in ie7, when displaying host & service problems the whole table
was displayd on top of the left menu making the whole page
looking all broken. by adding a height to the filter table this
problem is solved. it feels like a weird workaround. something
is seriously fishy about this. but it looks ok now so i'll leve
it like this for now.

Signed-off-by: My Appelgren <my@op5.com>