helpers/nagstat: Let nagstat::date_format() fetch date format
Instead of always parsing nagios.cfg to get the preferred date format,
we let the nagstat helper fetch it from nagios.cfg the first time every
user needs it and then store it in the database and as a session variable.
This means that nagios.cfg only has to be read once for each user and after
that only one db query has to be made each session to get this information.
Later versions will probably make this setting editable from the GUI on a
per user level.

Signed-off-by: Per Åsberg <per.asberg@op5.se>