models/host: Added cur_time to return UNIX_TIMESTAMP()
When we calculate duration, the value is produced by subtracting
last_state_change from current timestamp but since this will return
the actual timestamp if last_state_change is empty, we need some way
to detect if we should do any calculation on the duration value or not.
By adding cur_time with the current timestamp, we can compare to this
value when deciding.

Partial fix for issue #2519

Signed-off-by: Per Åsberg <per.asberg@op5.se>