helpers/config: Bypass get/set session value
Since session values can't be set by AJAX calls as long as we use a
value != 0 in config/session (because AJAX calls will generate a new
session ID), we need some way of handling dynamically modified values
as when updating page_refresh value.

By modifying the config helper to accept a fourth parameter - skip_session -
it is now possible to be sure that we always get the current value stored
in the database for a user.

Since it is only used for pag_refresh (in the refresh helper), the
overhead is negligible since it will only happen once for each page load.

Signed-off-by: Per Åsberg <per.asberg@op5.se>