template.php: Add possibility to disable refresh from controller
This patch gives us the possibility to disable the refresh functionality
from a controller. By adding $this->template->disable_refresh = true; to
any controller, the page won't refresh and the refresh control items will
be removed from the settings menu. If there's no widgets to display, the
entire settings menu with icon and all will be hidden. It will nevertheless
be shown if we have a page that has the "Hide header" functionality.

Signed-off-by: Per Åsberg <per.asberg@op5.se>