models/current_status: Skip parent join in fetch_hosts_causing_outages()
The sql query was making a join with host_parents table which made us
never get the hosts that we were really interested in.
This patch makes the method actually retur all hosts that are not
UP or DOWN and it's up to the other methods to figure out what to do with
it.

Partial fix for issue #2515


Signed-off-by: Per Åsberg <per.asberg@op5.se>