iframe-adjust.js: Use JQuery for iframe height
One big advantage when using JQuery instead of legacy
javascript code is that it is non obtrusive, meaning
that when for some reason an element isn't found on
the current page for some reason - it just ignores it
instead of throwing an error. In this case it was the
iframe element ID that wasn't found in some cases and
an ugly error would occur.

Signed-off-by: Per Åsberg <per.asberg@op5.se>