Don't always assume group data is available
When no host or servicegroups are defined the status controller
will throw an error when trying to view group_grid. This is because
the code assumes that there are always data.
This patch adds checks in the controller and makes sure that return
value from the models takes this into account.

Signed-off-by: Per Åsberg <per.asberg@op5.se>