collapse_menu.js: Suppress menu state error
When you click through the links in the menu, some AJAX functions
are unable to run because the page is unloading. This is perfectly
ok and having a javascript alert message that halts everything until
you click it is rather disturbing.
This patch simply wraps the alert in comments if we need to use it
later on.

Signed-off-by: Per Åsberg <per.asberg@op5.se>