Check return value before trying to iterate over db result
When fetching information on available {host,service}groups from db,
we lacked the appropriate checks that we actually have something to
iterate over. Also the group_grid template always assumed to get a
valid result to show.
This patch amends this by adding checks in the status controller
and in the template.

Signed-off-by: Per Åsberg <per.asberg@op5.se>