helpers/nagstat: Fix process_macros() to return something useful
Added support for 12 macros to start with. It ain't much but it's
a start.

the patch actually consists of two methods: process_macros() and
do_callback(). The latter one is because we sometimes need to make
the macro substitution using a callback of some kind. Currently we
support current_state substitution (host value 0 => 'UP', etc) and
state_type. The other macros are pretty straight forward using the
passed db object.

This is a partial fix for issue #2488

Signed-off-by: Per Åsberg <per.asberg@op5.se>