helpers/nagstat: Added helper stand alone helper function for status_text
As static method callbacks aren't yet available in our version of PHP
(5.1.6), we have to create a workaround for this using a regular function
that does the job for us. Once we have PHP 5.3.0 we can probably skip this
function since it should be possible as with any other function.

Signed-off-by: Per Åsberg <per.asberg@op5.se>