Added placeholder controllers for missing pages
Some pages are still missing and the underconstruction controller
doesn't handle calls that refer to specific pages (i.e links in
status/extinfo). Therefore these placeholder cotrollers are added
with the magic __call() method to handle all unknown methods. This
method redirects to the underconstruction controller to display the
info to the user.

Signed-off-by: Per Åsberg <per.asberg@op5.se>