models/ninja_setting: Make fetch_page_setting fetch default setting
Since we have both default and user settings for various things, we
sometimes want to be able to restore to the default settings.
This patch adds a third parameter, $default, which has a default
value of false but if set to true, it will return the default setting
and ignore the user preferences. It is then up to the calling method
to save this setting.

Signed-off-by: Per Åsberg <per.asberg@op5.se>