collapse_menu.js: Recalculate menu height on subsection collapse/expand
When menu sections are collapsed or expanded, we need to check if the
menu scroll/slider should be shown or not. Since there is a small delay
before the menu has been resized, we need to wait for 100msec before
doing the check and this is achieved by a setTimeout() call here.

Signed-off-by: Per Åsberg <per.asberg@op5.se>