controllers/ninja: Added add_template_path()
When adding resources in master template, we shouldn't have any hard
coded paths since this will make creating new templates rather cumbersome
since it will involve quite a lot of editing just to point to the correct
images, scripts and such. This method should be used together with add_path
to create the necessary paths pointing to the template selected in
config.php.

Signed-off-by: Per Åsberg <per.asberg@op5.se>