template.php: Don't hard code resource paths
This patch uses the new add_template_path() and add_path() to make it possible
to copy a theme to a new directory, change the current theme path on config.php
accordingly and then start to modify the theme without the need to edit
any paths in template.php.
The files in media/ are still linked in as before as these will be needed
in all templates.

Signed-off-by: Per Åsberg <per.asberg@op5.se>