controllers/ninja: Let ALLOW_PRODUCTION control Fire Profiler
Fire Profiler together with FirePHP is very useful to debug PHP code
and to print benchmarks for database queries etc. Having to switch it
on/off between commits is quite tedious so this commit makes the
ALLOW_PRODUCTION constant in the authenticated controller decide if
Fire Profiler should be enabled or not.

Signed-off-by: Per Åsberg <per.asberg@op5.se>