template.php: Make it possible to use links in breadcrumbs
We don't use links in breadcumbs anymore as default but sometimes it
might be a good idea to have this option. Aince $title is used
in <title> we must add a new variable when we want this.
This patch makes it possible to have an additional variable set by the
controller: $breadcrumb that can hold breadcrumb links.

Signed-off-by: Per Åsberg <per.asberg@op5.se>