controllers/status: Don't use Router::$method when we always need 'service'
The service method in the status controller is also called from
{host, service}group() and the $method will therefore vary depending
on what method was initially called. Hence we can't always assume that
we can rely on $method to use it as an argument to header_links since
that method can only handle 'host' or 'service' as first argument.
Failing to do so will result in some missing header fields in generated
table.

Signed-off-by: Per Åsberg <per.asberg@op5.se>