controllers/status: Don't add widget if no data
When no data is found we redirect to a separate template that displays
info to user. If we add any widget before this we will have a javascript
that assumes that the widget it "belongs to" exists. Hence, we add the
widget after checks against existing data has been made.

Signed-off-by: Per Åsberg <per.asberg@op5.se>