helpers/nagioscmd: s/host_name;service_description/service/
Since we select service by hostname;servicedescription, we'd otherwise
end up constructing commands where the hostname is printed twice. That
doesn't fly so well with Nagios, so we'd better avoid it. With this
patch, we only ever print the hostname once in the constructed command
that we write into the nagios pipe.

This probably fixes issue #2638.

Signed-off-by: Andreas Ericsson <ae@op5.se>