controllers/extinfo: Allow enabling of service eventhandlers
Previously there was a typo in the code causing the command to be
sent to Nagios to always be DISABLE_SVC_EVENT_HANDLER, which means
noone could ever enable event handlers for services via the ninja
gui.

This patch fixes that typo, allowing users to send the command
to enable service event handlers to Nagios.

Signed-off-by: Andreas Ericsson <ae@op5.se>