models/current_status: Active checks determine "disabledness"
The CGI's assume a host or service is "disabled" if active checks
are disabled, regardless of the state of passive checks. Ninja
did the same for services, but hosts had to have both active and
passive checks disabled for the host to truly show up as having
checks disabled.

While one can argue about the correctness of that, this patch at
least makes Ninja consistent with the CGI's.

This fixes issue #2621.

Signed-off-by: Andreas Ericsson <ae@op5.se>